refactor(repo C4): extract shared data types to engine/repository/types.ts
Moved ExpandedQuery, DocumentResult, SearchResult, RankedResult,
RRFContributionTrace, RRFScoreTrace, HybridQueryExplain, DocumentNotFound,
MultiGetResult, CollectionInfo, IndexStatus, IndexHealthInfo, and SnippetResult.
The barrel uses 'export type *' so these stay erased at runtime. Function-bearing
types (SearchHooks, HybridQueryOptions, etc.) remain in repository.ts pending
extraction of their callees.

Build: clean. Tests: 970/970 pass.
