+++ engine/repository.ts
@@ -113,10 +113,7 @@ export const RERANK_CANDIDATE_LIMIT = 40;
  * - vec: semantic variant → routes to vector only
  * - hyde: hypothetical document → routes to vector only
  */
-export type ExpandedQuery = {
-  type: 'lex' | 'vec' | 'hyde';
-  text: string;
-};
+// ExpandedQuery moved to engine/repository/types.ts (W1 C4).
 
 // =============================================================================
 // Path utilities — moved to engine/repository/paths.ts (W1 C2)
@@ -166,6 +163,21 @@ import {
   getActiveDocumentPaths,
 } from "./repository/content.js";
 import { indexSingleFile, unlinkSingleFile } from "./repository/indexing.js";
+import type {
+  ExpandedQuery,
+  DocumentResult,
+  SearchResult,
+  RankedResult,
+  RRFContributionTrace,
+  RRFScoreTrace,
+  HybridQueryExplain,
+  DocumentNotFound,
+  MultiGetResult,
+  CollectionInfo,
+  IndexStatus,
+  IndexHealthInfo,
+  SnippetResult,
+} from "./repository/types.js";
 
 // =============================================================================
 // Virtual Path Utilities (kindx://)
@@ -671,23 +683,7 @@ export function createStore(dbPath?: string): Store {
  * Unified document result type with all metadata.
  * Body is optional - use getDocumentBody() to load it separately if needed.
  */
-export type DocumentResult = {
-  filepath: string;           // Full filesystem path
-  displayPath: string;        // Short display path (e.g., "docs/readme.md")
-  title: string;              // Document title (from first heading or filename)
-  context: string | null;     // Folder context description if configured
-  hash: string;               // Content hash for caching/change detection
-  docid: string;              // Short docid (first 6 chars of hash) for quick reference
-  collectionName: string;     // Parent collection name
-  modifiedAt: string;         // Last modification timestamp
-  bodyLength: number;         // Body length in bytes (useful before loading)
-  body?: string;              // Document body (optional, load with getDocumentBody)
-  extraction?: {
-    format: string;
-    extractor: string;
-    warnings: string[];
-  };
-};
+// DocumentResult moved to engine/repository/types.ts (W1 C4).
 
 /**
  * Extract short docid from a full hash (first 6 characters).
@@ -767,122 +763,9 @@ export function handelize(path: string): string {
   return result;
 }
 
-/**
- * Search result extends DocumentResult with score and source info
- */
-export type SearchResult = DocumentResult & {
-  score: number;              // Relevance score (0-1)
-  source: "fts" | "vec";      // Search source (full-text or vector)
-  chunkPos?: number;          // Character position of matching chunk (for vector search)
-};
-
-/**
- * Ranked result for RRF fusion (simplified, used internally)
- */
-export type RankedResult = {
-  file: string;
-  displayPath: string;
-  title: string;
-  body: string;
-  score: number;
-  modifiedAt?: string;
-};
-
-export type RRFContributionTrace = {
-  listIndex: number;
-  source: "fts" | "vec";
-  queryType: "original" | "lex" | "vec" | "hyde";
-  query: string;
-  rank: number;            // 1-indexed rank within list
-  weight: number;
-  backendScore: number;    // Backend-normalized score before fusion
-  rrfContribution: number; // weight / (k + rank)
-};
-
-export type RRFScoreTrace = {
-  contributions: RRFContributionTrace[];
-  baseScore: number;       // Sum of reciprocal-rank contributions
-  topRank: number;         // Best (lowest) rank seen across lists
-  topRankBonus: number;    // +0.05 for rank 1, +0.02 for rank 2-3
-  totalScore: number;      // baseScore + topRankBonus
-};
-
-export type HybridQueryExplain = {
-  ftsScores: number[];
-  vectorScores: number[];
-  rrf: {
-    rank: number;          // Rank after RRF fusion (1-indexed)
-    positionScore: number; // 1 / rank used in position-aware blending
-    weight: number;        // Position-aware RRF weight (0.75 / 0.60 / 0.40)
-    baseScore: number;
-    topRankBonus: number;
-    totalScore: number;
-    contributions: RRFContributionTrace[];
-  };
-  rerankScore: number;
-  blendedScore: number;
-};
-
-/**
- * Error result when document is not found
- */
-export type DocumentNotFound = {
-  error: "not_found";
-  query: string;
-  similarFiles: string[];
-};
-
-/**
- * Result from multi-get operations
- */
-export type MultiGetResult = {
-  doc: DocumentResult;
-  skipped: false;
-} | {
-  doc: Pick<DocumentResult, "filepath" | "displayPath">;
-  skipped: true;
-  skipReason: string;
-};
-
-export type CollectionInfo = {
-  name: string;
-  path: string;
-  pattern: string;
-  documents: number;
-  lastUpdated: string;
-};
-
-export type IndexStatus = {
-  totalDocuments: number;
-  needsEmbedding: number;
-  hasVectorIndex: boolean;
-  capabilities: Record<string, string>;
-  ann: {
-    enabled: boolean;
-    mode: "ann" | "exact";
-    state: "ready" | "stale" | "missing" | "degraded";
-    probeCount: number;
-    shortlistLimit: number;
-    details: Array<{ collection: string; shard: number; state: "ready" | "stale" | "missing" | "degraded"; reason: string }>;
-  };
-  encryption: {
-    encrypted: boolean;
-    keyConfigured: boolean;
-    bytes: number;
-  };
-  ingestion: {
-    warnedDocuments: number;
-    byFormat: Array<{ format: string; count: number }>;
-    byWarning: Array<{ warning: string; count: number }>;
-  };
-  collections: CollectionInfo[];
-  shards: {
-    enabledCollections: Array<{ collection: string; shardCount: number }>;
-    checkpointPath: string;
-    checkpointExists: boolean;
-    warnings: string[];
-  };
-};
+// SearchResult, RankedResult, RRFContributionTrace, RRFScoreTrace, HybridQueryExplain,
+// DocumentNotFound, MultiGetResult, CollectionInfo, IndexStatus moved to
+// engine/repository/types.ts (W1 C4).
 
 // =============================================================================
 // Index health
@@ -898,11 +781,7 @@ export function getHashesNeedingEmbedding(db: Database): number {
   return result.count;
 }
 
-export type IndexHealthInfo = {
-  needsEmbedding: number;
-  totalDocs: number;
-  daysStale: number | null;
-};
+// IndexHealthInfo moved to engine/repository/types.ts (W1 C4).
 
 export function getIndexHealth(db: Database): IndexHealthInfo {
   const needsEmbedding = getHashesNeedingEmbedding(db);
@@ -2336,13 +2215,7 @@ export function getStatus(db: Database): IndexStatus {
 // Snippet extraction
 // =============================================================================
 
-export type SnippetResult = {
-  line: number;           // 1-indexed line number of best match
-  snippet: string;        // The snippet text with diff-style header
-  linesBefore: number;    // Lines in document before snippet
-  linesAfter: number;     // Lines in document after snippet
-  snippetLines: number;   // Number of lines in snippet
-};
+// SnippetResult moved to engine/repository/types.ts (W1 C4).
 
 export function extractSnippet(body: string, query: string, maxLen = 500, chunkPos?: number, chunkLen?: number): SnippetResult {
   const totalLines = body.split('\n').length;
